+++ src/collective/upload/testing.py
@@ -9,20 +9,19 @@ plone.app.contenttypes:
 from plone.app.robotframework.testing import AUTOLOGIN_LIBRARY_FIXTURE
 from plone.app.testing import FunctionalTesting
 from plone.app.testing import IntegrationTesting
-from plone.app.testing import PLONE_FIXTURE
 from plone.app.testing import PloneSandboxLayer
 from plone.testing import z2
 
-import os
 import pkg_resources
 
 try:
     pkg_resources.get_distribution('plone.app.contenttypes')
 except pkg_resources.DistributionNotFound:
+    from plone.app.testing import PLONE_FIXTURE
     DEXTERITY_ONLY = False
 else:
-    # this environment variable is set in .travis.yml test matrix
-    DEXTERITY_ONLY = os.environ.get('DEXTERITY_ONLY') is not None
+    from plone.app.contenttypes.testing import PLONE_APP_CONTENTTYPES_FIXTURE as PLONE_FIXTURE
+    DEXTERITY_ONLY = True
 
 
 class Fixture(PloneSandboxLayer):
@@ -30,18 +29,10 @@ class Fixture(PloneSandboxLayer):
     defaultBases = (PLONE_FIXTURE,)
 
     def setUpZope(self, app, configurationContext):
-        if DEXTERITY_ONLY:
-            import plone.app.contenttypes
-            self.loadZCML(package=plone.app.contenttypes)
-            z2.installProduct(app, 'Products.DateRecurringIndex')
-
         import collective.upload
         self.loadZCML(package=collective.upload)
 
     def setUpPloneSite(self, portal):
-        if DEXTERITY_ONLY:
-            self.applyProfile(portal, 'plone.app.contenttypes:default')
-
         self.applyProfile(portal, 'collective.upload:default')
 
 FIXTURE = Fixture()