+++ src/collective/upload/tests/test_controlpanel.py
@@ -1,17 +1,18 @@
 # -*- coding: utf-8 -*-
 from collective.upload import config
 from collective.upload.config import PROJECTNAME
+from collective.upload.interfaces import IUploadBrowserLayer
 from collective.upload.interfaces import IUploadSettings
 from collective.upload.testing import INTEGRATION_TESTING
+from plone import api
 from plone.app.testing import logout
 from plone.registry.interfaces import IRegistry
-from zope.component import getMultiAdapter
 from zope.component import getUtility
+from zope.interface import alsoProvides
 
 import unittest
 
-
-BASE_REGISTRY = 'collective.upload.controlpanel.IuploadSettings.%s'
+BASE_REGISTRY = 'collective.upload.controlpanel.IUploadSettings.'
 
 
 class ControlPanelTestCase(unittest.TestCase):
@@ -20,20 +21,21 @@ class ControlPanelTestCase(unittest.TestCase):
 
     def setUp(self):
         self.portal = self.layer['portal']
+        self.request = self.layer['request']
+        alsoProvides(self.request, IUploadBrowserLayer)
         self.controlpanel = self.portal['portal_controlpanel']
 
     def test_controlpanel_has_view(self):
-        view = getMultiAdapter((self.portal, self.portal.REQUEST),
-                               name='upload-settings')
+        context, request = self.portal, self.request
+        view = api.content.get_view('upload-settings', context, request)
         view = view.__of__(self.portal)
         self.assertTrue(view())
 
     def test_controlpanel_view_is_protected(self):
         from AccessControl import Unauthorized
         logout()
-        self.assertRaises(Unauthorized,
-                          self.portal.restrictedTraverse,
-                          '@@upload-settings')
+        with self.assertRaises(Unauthorized):
+            self.portal.restrictedTraverse('@@upload-settings')
 
     def test_controlpanel_installed(self):
         actions = [a.getAction(self)['id']
@@ -46,6 +48,14 @@ class ControlPanelTestCase(unittest.TestCase):
         actions = [a.getAction(self)['id']
                    for a in self.controlpanel.listActions()]
         self.assertNotIn('upload', actions)
+
+    def test_controlpanel_permissions(self):
+        roles = ['Manager', 'Site Administrator']
+        for r in roles:
+            with api.env.adopt_roles([r]):
+                configlets = self.controlpanel.enumConfiglets(group='Products')
+                configlets = [a['id'] for a in configlets]
+                self.assertIn('upload', configlets, 'not listed: ' + r)
 
 
 class RegistryTestCase(unittest.TestCase):
@@ -86,11 +96,11 @@ class RegistryTestCase(unittest.TestCase):
         qi.uninstallProducts(products=[config.PROJECTNAME])
 
         records = [
-            BASE_REGISTRY % 'show_widget',
-            BASE_REGISTRY % 'upload_extensions',
-            BASE_REGISTRY % 'max_file_size',
-            BASE_REGISTRY % 'resize_max_width',
-            BASE_REGISTRY % 'resize_max_height',
+            BASE_REGISTRY + 'show_widget',
+            BASE_REGISTRY + 'upload_extensions',
+            BASE_REGISTRY + 'max_file_size',
+            BASE_REGISTRY + 'resize_max_width',
+            BASE_REGISTRY + 'resize_max_height',
         ]
 
         for r in records: